requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
boto3==1.34.0
pandas==2.2.0
scikit-learn==1.4.0
httpx==0.26.0
numpy==1.26.4

langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
langchain-groq>=0.1.3
llama-cpp-python